Refactor audit hotspot modules
