--- apps/desktop/src/main/custom-skills.ts
@@ -1,7 +1,0 @@
-export {
-  getCustomSkill,
-  listCustomSkills,
-  readSkillBundleDirectory,
-  removeCustomSkill,
-  saveCustomSkill,
-} from './native-customizations.ts'
