+++ apps/desktop/src/main/ipc-handlers.ts
@@ -1,8 +1,6 @@
 import type { BrowserWindow, IpcMain, IpcMainEvent, IpcMainInvokeEvent } from 'electron'
 import type {
-  CapabilityTool,
   CapabilityToolEntry,
-  CustomAgentConfig,
   DestructiveConfirmationRequest,
   RuntimeContextOptions,
   ScopedArtifactRef,
@@ -10,9 +8,6 @@ import type {
   ToolListOptions,
 } from '@open-cowork/shared'
 import { isMcpAuthRequiredStatus } from '@open-cowork/shared'
-import { Client as McpClient } from '@modelcontextprotocol/sdk/client/index'
-import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio'
-import { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp'
 import { dirname, join, resolve } from 'path'
 import { fileURLToPath } from 'url'
 import { getChartArtifactsRoot } from './chart-artifacts.ts'
@@ -28,23 +23,14 @@ import { log } from './logger.ts'
 import { getMcpStatus } from './events.ts'
 import { shortSessionId } from './log-sanitizer.ts'
 import { dispatchRuntimeSessionEvent, setSessionHistoryRefreshHandler } from './session-event-dispatcher.ts'
-import { getCustomAgentCatalog } from './custom-agents.ts'
-import { listBuiltInAgentDetails } from './agent-config.ts'
 import { getSessionRecord, listSessionRecords } from './session-registry.ts'
 import { syncSessionView } from './session-history-loader.ts'
 import { ensureRuntimeContextDirectory } from './runtime-context.ts'
-import { humanizeToolId, isVisibleRuntimeToolId, runtimeToolId } from './runtime-tools.ts'
-import { validateCustomMcpStdioCommand } from './mcp-stdio-policy.ts'
+import { isVisibleRuntimeToolId, runtimeToolId } from './runtime-tools.ts'
 import { createSandboxWorkspaceDir, isSandboxWorkspaceDir } from './runtime-paths.ts'
-import { listCustomMcps } from './native-customizations.ts'
 import { createDestructiveConfirmationManager } from './destructive-actions.ts'
 import { sessionEngine } from './session-engine.ts'
 import { listAutomationState } from './automation-store.ts'
-import {
-  type ResolvedRuntimeMcpEntry,
-  resolveConfiguredMcpRuntimeEntry,
-  resolveCustomMcpRuntimeEntryForRuntime,
-} from './runtime-mcp.ts'
 import { observePerf } from './perf-metrics.ts'
 import { registerAppHandlers } from './ipc/app-handlers.ts'
 import { registerArtifactHandlers } from './ipc/artifact-handlers.ts'
@@ -58,6 +44,13 @@ import { clearPermissionsForSession, trackPermission } from './permission-tracke
 import { ProjectDirectoryGrantRegistry, trustedRecordDirectoryMatches } from './directory-grants.ts'
 import { resolveContainedArtifactPath } from './artifact-path-policy.ts'
 import { isTrustedRendererIpcUrl } from './main-window-lifecycle.ts'
+import { delay } from './delay.ts'
+import {
+  buildCustomAgentPermission,
+  createCapabilityToolDiscovery,
+  isLikelyMcpAuthError,
+  listToolsFromMcpEntry,
+} from './capability-tool-discovery.ts'
 
 import { RUNTIME_TOOL_CACHE_TTL_MS, runtimeToolCache } from './runtime-tool-cache.ts'
 export { invalidateRuntimeToolCache } from './runtime-tool-cache.ts'
@@ -219,29 +212,6 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
     return { ...target, directory: null }
   }
 
-  async function buildCustomAgentPermission(agent: CustomAgentConfig, options?: RuntimeContextOptions) {
-    const catalog = await getCustomAgentCatalog(options)
-    const selectedTools = catalog.tools.filter((tool) => agent.toolIds.includes(tool.id))
-    const allowPatterns = Array.from(new Set(selectedTools.flatMap((tool) => tool.allowPatterns)))
-    const askPatterns = Array.from(new Set(selectedTools.flatMap((tool) => tool.askPatterns)))
-    const deniedPatterns = Array.from(new Set((agent.deniedToolPatterns || []).map((pattern) => pattern.trim()).filter(Boolean)))
-
-    const permission: Record<string, unknown> = {}
-    if ((agent.skillNames || []).length > 0) {
-      permission.skill = Object.fromEntries((agent.skillNames || []).map((name) => [name, 'allow']))
-    }
-
-    for (const pattern of allowPatterns) permission[pattern] = 'allow'
-    for (const pattern of askPatterns) permission[pattern] = 'ask'
-    // Specific user-chosen denies land LAST so they shadow the MCP's
-    // wildcard allow when the same key is written (e.g. a user denies
-    // `mcp__github__*` outright). OpenCode's permission resolver picks
-    // the most specific match, so patterns like `mcp__github__delete_repo`
-    // coexist with `mcp__github__*: allow` without key collision.
-    for (const pattern of deniedPatterns) permission[pattern] = 'deny'
-    return permission
-  }
-
   function logHandlerError(handler: string, err: unknown) {
     const message = err instanceof Error
       ? err.message
@@ -310,79 +280,6 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
     return { client, record, directory }
   }
 
-  function capabilityToolPrefixes(tool: CapabilityTool) {
-    const prefixes = new Set<string>()
-
-    if (tool.namespace) {
-      prefixes.add(`mcp__${tool.namespace}__`)
-      prefixes.add(`${tool.namespace}_`)
-    }
-
-    prefixes.add(`mcp__${tool.id}__`)
-    prefixes.add(`${tool.id}_`)
-
-    return Array.from(prefixes)
-  }
-
-  function runtimeToolMatchesCapability(entry: unknown, tool: CapabilityTool) {
-    const id = runtimeToolId(entry)
-    if (!id) return false
-    if (id === tool.id) return true
-    return capabilityToolPrefixes(tool).some((prefix) => id.startsWith(prefix))
-  }
-
-  function isMcpBackedCapability(tool: CapabilityTool) {
-    return Boolean(tool.namespace) || tool.patterns.some((pattern) => pattern.startsWith('mcp__'))
-  }
-
-  async function listToolsFromMcpEntry(entry: unknown) {
-    if (!entry) return []
-
-    const runtimeEntry = entry as ResolvedRuntimeMcpEntry
-    const client = new McpClient(
-      { name: 'open-cowork-capabilities', version: '1.0.0' },
-      { capabilities: {} },
-    )
-
-    if (runtimeEntry.type === 'local') {
-      const [command, ...args] = runtimeEntry.command
-      if (!command) return []
-      const transport = new StdioClientTransport({
-        command,
-        args,
-        env: runtimeEntry.environment,
-        stderr: 'pipe',
-      })
-      await client.connect(transport)
-    } else {
-      const transport = new StreamableHTTPClientTransport(new URL(runtimeEntry.url), {
-        requestInit: runtimeEntry.headers
-          ? { headers: runtimeEntry.headers }
-          : undefined,
-      })
-      await client.connect(transport)
-    }
-
-    try {
-      const result = await client.listTools()
-      return (result.tools || []).map((tool: { name: string; description?: string }) => ({
-        id: tool.name,
-        description: tool.description?.trim() || 'No description available for this MCP method.',
-      }))
-    } finally {
-      await client.close().catch(() => {})
-    }
-  }
-
-  function isLikelyMcpAuthError(error: unknown) {
-    const message = error instanceof Error ? error.message : String(error || '')
-    return /missing authorization header|invalid_token|unauthorized|forbidden|40[13]|needs_auth|oauth/i.test(message)
-  }
-
-  function wait(ms: number) {
-    return new Promise((resolvePromise) => setTimeout(resolvePromise, ms))
-  }
-
   async function waitForMcpStatus(name: string, timeoutMs = 10_000) {
     const deadline = Date.now() + timeoutMs
     while (Date.now() < deadline) {
@@ -392,7 +289,7 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
         const match = statuses.find((entry) => entry.name === name)
         if (match) return match
       }
-      await wait(500)
+      await delay(500)
     }
     return null
   }
@@ -412,65 +309,6 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
     } catch (error) {
       logHandlerError(`custom:add-mcp auth ${name}`, error)
     }
-  }
-
-  async function discoverCapabilityToolEntries(tool: CapabilityTool, options?: RuntimeContextOptions) {
-    if (!isMcpBackedCapability(tool)) return tool.availableTools || []
-
-    const cacheKey = `${tool.source}:${tool.id}:${tool.namespace || ''}:${resolveContextDirectory(options) || 'machine'}`
-    const cached = capabilityToolMethodCache.get(cacheKey)
-    if (cached && cached.expiresAt > Date.now()) {
-      return cached.entries
-    }
-
-    const settings = getEffectiveSettings()
-    const builtinEntry = tool.source === 'builtin'
-      ? resolveConfiguredMcpRuntimeEntry(tool.namespace || tool.id, settings)
-      : null
-    const matchingCustomMcp = tool.source === 'custom'
-      ? listCustomMcps(options).find((entry) => entry.name === tool.id || entry.name === tool.namespace) || null
-      : null
-    const customEntry = matchingCustomMcp
-      ? await resolveCustomMcpRuntimeEntryForRuntime(matchingCustomMcp)
-      : null
-
-    if (matchingCustomMcp?.type === 'stdio') {
-      try {
-        validateCustomMcpStdioCommand(matchingCustomMcp)
-      } catch (error) {
-        logHandlerError(`capability:mcp-tools ${tool.id}`, error)
-        return []
-      }
-    }
-
-    const shouldSkipDirectProbe = Boolean(
-      matchingCustomMcp
-      && matchingCustomMcp.type === 'http'
-      && (!matchingCustomMcp.headers || Object.keys(matchingCustomMcp.headers).length === 0),
-    )
-    if (shouldSkipDirectProbe) return []
-
-    const entries = await listToolsFromMcpEntry(builtinEntry || customEntry).catch((error) => {
-      logHandlerError(`capability:mcp-tools ${tool.id}`, error)
-      return []
-    })
-
-    // Bound the cache so a long-running session with many projects × tools
-    // can't grow the map unbounded. The key space is
-    // `source:id:namespace:directory` so collisions across users are rare;
-    // when we hit the cap we evict the oldest entry (FIFO is sufficient
-    // since all entries share the same 30s TTL).
-    const CAPABILITY_CACHE_MAX = 500
-    if (capabilityToolMethodCache.size >= CAPABILITY_CACHE_MAX) {
-      const oldestKey = capabilityToolMethodCache.keys().next().value
-      if (oldestKey !== undefined) capabilityToolMethodCache.delete(oldestKey)
-    }
-    capabilityToolMethodCache.set(cacheKey, {
-      expiresAt: Date.now() + 30_000,
-      entries,
-    })
-
-    return entries
   }
 
   async function listRuntimeTools(options?: ToolListOptions) {
@@ -519,82 +357,11 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
     }
   }
 
-
-  async function withDiscoveredBuiltInTools(tools: CapabilityTool[], runtimeTools: unknown[], options?: RuntimeContextOptions & { deep?: boolean }) {
-    const builtInAgentDetails = listBuiltInAgentDetails()
-    const nativeToolEntries = new Map<string, CapabilityTool>()
-
-    for (const entry of runtimeTools) {
-      const id = runtimeToolId(entry)
-      if (!id) continue
-      if (!isVisibleRuntimeToolId(id)) continue
-      if (id.startsWith('mcp__')) continue
-      if (tools.some((tool) => runtimeToolMatchesCapability(entry, tool))) continue
-
-      const agentNames = builtInAgentDetails
-        .filter((agent) => agent.nativeToolIds.includes(id))
-        .map((agent) => agent.label)
-
-      nativeToolEntries.set(id, {
-        id,
-        name: humanizeToolId(id),
-        description: typeof (entry as { description?: string })?.description === 'string' && (entry as { description?: string }).description!.trim().length > 0
-          ? (entry as { description?: string }).description!.trim()
-          : 'Native OpenCode tool available in the current runtime context.',
-        kind: 'built-in',
-        source: 'builtin',
-        origin: 'opencode',
-        namespace: null,
-        patterns: [id],
-        availableTools: [
-          {
-            id,
-            description: typeof (entry as { description?: string })?.description === 'string' && (entry as { description?: string }).description!.trim().length > 0
-              ? (entry as { description?: string }).description!.trim()
-              : 'Native OpenCode tool available in the current runtime context.',
-          },
-        ],
-        agentNames,
-      })
-    }
-
-    const combined = [...tools, ...nativeToolEntries.values()].sort((a, b) => a.name.localeCompare(b.name))
-
-    return Promise.all(combined.map(async (tool) => {
-      const runtimeEntries = runtimeTools
-        .filter((entry) => runtimeToolMatchesCapability(entry, tool))
-        .map((entry) => ({
-          id: runtimeToolId(entry),
-          description: typeof (entry as { description?: string })?.description === 'string' && (entry as { description?: string }).description!.trim().length > 0
-            ? (entry as { description?: string }).description!.trim()
-            : 'No description available for this MCP method.',
-        }))
-        .filter((entry): entry is CapabilityToolEntry => Boolean(entry.id))
-
-      if (runtimeEntries.length > 0) {
-        return { ...tool, availableTools: runtimeEntries }
-      }
-
-      if (!isMcpBackedCapability(tool)) {
-        return tool
-      }
-
-      // `deep` is the opt-in for the expensive per-MCP probe. List
-      // views (the Capabilities grid) skip it: the card renders from
-      // name + description + icon alone, and probing 16 MCPs on every
-      // page open pushes the IPC into 3-5s. Detail views
-      // (capabilities.tool(id)) pass deep:true so the method table is
-      // populated when the user actually opens one tool.
-      if (!options?.deep) return tool
-
-      const fallbackEntries = await discoverCapabilityToolEntries(tool, options)
-      if (fallbackEntries.length > 0) {
-        return { ...tool, availableTools: fallbackEntries }
-      }
-
-      return tool
-    }))
-  }
+  const capabilityToolDiscovery = createCapabilityToolDiscovery({
+    resolveContextDirectory,
+    logHandlerError,
+    capabilityToolMethodCache,
+  })
 
   const context: IpcHandlerContext = {
     ipcMain: instrumentedIpcMain,
@@ -614,7 +381,7 @@ export function setupIpcHandlers(ipcMain: IpcMain, getMainWindow: () => BrowserW
     getSessionClient,
     getSessionV2Client,
     listRuntimeTools,
-    withDiscoveredBuiltInTools,
+    withDiscoveredBuiltInTools: capabilityToolDiscovery.withDiscoveredBuiltInTools,
     listToolsFromMcpEntry,
     isLikelyMcpAuthError,
     authenticateNewRemoteMcpIfNeeded,
