+++ apps/desktop/src/renderer/components/capabilities/CapabilitiesPage.tsx
@@ -9,69 +9,18 @@ import { confirmMcpRemoval, confirmSkillRemoval } from '../../helpers/destructiv
 import { SkillSelectionCard, ToolSelectionCard } from './CapabilitySelectionCard'
 import { getBrandName } from '../../helpers/brand'
 import { t } from '../../helpers/i18n'
-
-type Tab = 'tools' | 'skills'
-type Selection =
-  | { type: 'tool'; id: string }
-  | { type: 'skill'; name: string }
-  | null
-
-function stripFrontmatter(content: string) {
-  return content.replace(/^---[\s\S]*?---\n?/, '').trim()
-}
-
-function prettyKind(tool: CapabilityTool) {
-  if (tool.origin === 'opencode') return t('capabilities.kindOpencodeTool', 'OpenCode tool')
-  if (tool.source === 'custom') return t('capabilities.kindCustomMcp', 'Custom MCP')
-  return tool.kind === 'built-in' ? t('capabilities.kindBuiltinTool', 'Built-in tool') : t('capabilities.kindMcpTool', 'MCP tool')
-}
-
-function prettySkillKind(skill: CapabilitySkill) {
-  if (skill.source === 'custom') return t('capabilities.kindCustomSkill', 'Custom skill')
-  return t('capabilities.kindBuiltinSkill', 'Built-in skill')
-}
-
-function prettySkillSource(skill: CapabilitySkill) {
-  if (skill.origin === 'open-cowork') return t('capabilities.skillSourceBundled', '{{brand}} bundled skill', { brand: getBrandName() })
-  if (skill.scope === 'project') return t('capabilities.skillSourceProject', 'Project skill')
-  if (skill.scope === 'machine') return t('capabilities.skillSourceMachine', 'Machine skill')
-  return t('capabilities.skillSourceBundle', 'Skill bundle')
-}
-
-function toolPrefixes(tool: CapabilityTool) {
-  const prefixes = new Set<string>()
-
-  if (tool.namespace) {
-    prefixes.add(`mcp__${tool.namespace}__`)
-    prefixes.add(`${tool.namespace}_`)
-  }
-
-  prefixes.add(`mcp__${tool.id}__`)
-  prefixes.add(`${tool.id}_`)
-
-  return Array.from(prefixes)
-}
-
-function safeText(value: string | null | undefined) {
-  return typeof value === 'string' ? value : ''
-}
-
-function mergedRuntimeToolset(tool: CapabilityTool, runtimeTools: RuntimeToolDescriptor[]) {
-  const prefixes = toolPrefixes(tool)
-  const discovered = runtimeTools.filter((entry) => {
-    const id = entry.id || entry.name || ''
-    return id === tool.id || prefixes.some((prefix) => id.startsWith(prefix))
-  })
-
-  if (discovered.length > 0) {
-    return discovered.map((entry) => ({
-      id: entry.id || entry.name || 'unknown',
-      description: entry.description || 'No description available for this MCP method.',
-    }))
-  }
-
-  return tool.availableTools || []
-}
+import {
+  buildAgentSeedFromSkill,
+  buildAgentSeedFromTool,
+  mergedRuntimeToolset,
+  prettyKind,
+  prettySkillKind,
+  prettySkillSource,
+  safeText,
+  stripFrontmatter,
+  type Selection,
+  type Tab,
+} from './capabilities-page-support.ts'
 
 function StatBox({ label, value }: { label: string; value: string }) {
   return (
@@ -464,34 +413,6 @@ function ToolIntegrationToggleCard({
       ) : null}
     </div>
   )
-}
-
-function suggestAgentId(value: string) {
-  return `${value.trim().toLowerCase().replace(/[^a-z0-9]+/g, '-').replace(/^-+|-+$/g, '') || 'new'}-agent`
-}
-
-function buildAgentSeedFromTool(tool: CapabilityTool): Partial<CustomAgentConfig> {
-  return {
-    name: suggestAgentId(tool.id),
-    description: tool.description,
-    toolIds: [tool.id],
-    instructions: '',
-    skillNames: [],
-    enabled: true,
-    color: 'accent',
-  }
-}
-
-function buildAgentSeedFromSkill(skill: CapabilitySkill): Partial<CustomAgentConfig> {
-  return {
-    name: suggestAgentId(skill.name),
-    description: skill.description,
-    toolIds: [...(skill.toolIds || [])],
-    instructions: '',
-    skillNames: [skill.name],
-    enabled: true,
-    color: 'accent',
-  }
 }
 
 export function CapabilitiesPage({
