+++ tests/custom-skills.test.ts
@@ -2,7 +2,7 @@ import test from 'node:test'
 import assert from 'node:assert/strict'
 import { mkdtempSync, mkdirSync, rmSync, writeFileSync } from 'fs'
 import { join } from 'path'
-import { listCustomSkills, readSkillBundleDirectory, saveCustomSkill } from '../apps/desktop/src/main/custom-skills.ts'
+import { listCustomSkills, readSkillBundleDirectory, saveCustomSkill } from '../apps/desktop/src/main/native-customizations.ts'
 import { CUSTOM_SKILL_LIMITS } from '../apps/desktop/src/main/custom-content-limits.ts'
 import { clearConfigCaches } from '../apps/desktop/src/main/config-loader.ts'
 import { getMachineSkillsDir } from '../apps/desktop/src/main/runtime-paths.ts'
